requirements.txt
deepdiff==6.3.0
gymnasium==0.28.1
holidays==0.21.13
matplotlib==3.7.1
numpy==1.24.3
openpyxl==3.0.10
pandas==1.5.2
pathlib==1.0.1
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2022.7
PyYAML==6.0
stable-baselines3==2.2.1
torch==2.2.0
wandb==0.16.1